refactor: split config helpers for story 14.3
